Swagger documentation + minor fix v2.0

+++ routes/transaction.route.js
@@ -11,14 +11,18 @@ import { transactionViewModel } from '../view_models/transaction.viewModel.js'
 import { v4 as uuidv4 } from 'uuid'
 import debtReminderModel from '../models/debtReminder.model.js'
 import { debtReminderViewModel } from '../view_models/debtReminder.viewModel.js'
+import { transactionOTPViewModel } from '../view_models/transactionOTP.viewModel.js'
 
 const router = express.Router()
 
 router.use(currentUserMdw)
 router.post('/:id/ResendOTP', async (req, res) => {
   const currentUser = res.locals.currentUser
+  if (currentUser.role_id !== 2) {
+    return res.status(403).json({ message: 'You do not have permission to access the API!' })
+  }
   const { id } = req.params
-  const transaction = await transactionModel.findOne({ id }, transactionViewModel.split(' '))
+  const transaction = await transactionModel.findOne({ id }, transactionViewModel)
   const transporter = nodemailer.createTransport({
     service: 'Gmail',
     auth: {
@@ -53,34 +57,37 @@ router.post('/:id/ResendOTP', async (req, res) => {
           transaction_id: id
         }
         await transactionOTPModel.add(otpData)
-        return res.status(200).json({ message: 'Gửi mail thành công' })
+        return res.status(201).json({ message: 'New OTP is sent' })
       }
     })
   } catch (err) {
     console.log('err', err)
-    return res.status(200).json({ message: 'Gửi mail thất bại' })
+    return res.status(200).json({ message: 'Failed to sent email' })
   }
 })
 
 router.post('/VerifyOTP', async (req, res) => {
   const currentUser = res.locals.currentUser
+  if (currentUser.role_id !== 2) {
+    return res.status(403).json({ message: 'You do not have permission to access the API!' })
+  }
   const data = req.body
-  const otp = await transactionOTPModel.findOne({ otp: data.otp }, transactionViewModel)
+  const otp = await transactionOTPModel.findOne({ otp: data.otp }, transactionOTPViewModel)
   if (!otp) {
-    return res.status(200).json({ error_message: 'OTP không chính xác' })
+    return res.status(200).json({ message: 'Invalid OTP' })
   }
   if (new Date(otp.expired_at).getTime() < new Date().getTime()) {
     await transactionOTPModel.delete(otp.id)
-    return res.status(200).json({ error_message: 'OTP hết hạn' })
+    return res.status(200).json({ message: 'OTP expired' })
   }
   let transactionData = await transactionModel.findOne({ id: otp.transaction_id }, transactionViewModel)
   const sourceAccount = await accountModel.findOne({ uuid: transactionData.source_account_uuid, user_id: currentUser.id }, accountViewModel)
   if (!sourceAccount) {
-    return res.status(200).json({ error_message: 'Không tìm thấy tài khoản thanh toán' })
+    return res.status(200).json({ message: 'Cannot find payment account' })
   }
-  const destinationAccount = await accountModel.fetch({ uuid: transactionData.destination_account_uuid }, accountViewModel)
+  const destinationAccount = (await accountModel.fetch({ uuid: transactionData.destination_account_uuid }, accountViewModel))[0]
   if (!destinationAccount) {
-    return res.status(200).json({ error_message: 'Không tìm thấy tài khoản người nhận' })
+    return res.status(200).json({ message: 'Cannot find receiver account' })
   }
   transactionData = {
     ...transactionData,
@@ -89,38 +96,44 @@ router.post('/VerifyOTP', async (req, res) => {
   await transactionModel.update(transactionData.id, transactionData)
 
   if (sourceAccount.balance < transactionData.amount) {
-    return res.status(200).json({ error_message: 'Không đủ số dư' })
+    return res.status(200).json({ message: 'The balance is not enough' })
   }
   sourceAccount.balance -= transactionData.amount
   destinationAccount.balance += transactionData.amount
 
   if (transactionData.debt_reminder_id !== null) {
-    let debtReminder = debtReminderModel.findOne({ id: transactionData.debt_reminder_id }, debtReminderViewModel)
+    let debtReminder = await debtReminderModel.findOne({ id: transactionData.debt_reminder_id }, debtReminderViewModel)
     if (!debtReminder) {
-      return res.status(200).json({ error_message: 'Không tìm thấy nhắc nợ' })
+      return res.status(200).json({ message: 'Cannot find Debt reminder' })
     }
     debtReminder = {
       ...debtReminder,
       isPaid: true
     }
+    console.log('debtReminder: ', debtReminder);
     await debtReminderModel.update(debtReminder.id, debtReminder)
   }
+  console.log('sourceAccount: ', sourceAccount);
   await accountModel.update(sourceAccount.id, sourceAccount)
+  console.log('destinationAccount: ', destinationAccount);
   await accountModel.update(destinationAccount.id, destinationAccount)
   await transactionOTPModel.delete(otp.id)
-  res.status(200).json({ message: 'Xác nhận thành công' })
+  res.status(201).json({ message: 'OTP code confirm succesfully. Transaction has been proceeded' })
 })
 
 router.post('/', async (req, res) => {
   const currentUser = res.locals.currentUser
+  if (currentUser.role_id !== 2) {
+    return res.status(403).json({ message: 'You do not have permission to access the API!' })
+  }
   const data = req.body
   const sourceAccount = await accountModel.findOne({ uuid: data.source_account_uuid, user_id: currentUser.id }, accountViewModel)
   if (!sourceAccount) {
-    return res.status(200).json({ message: 'Không tìm thấy tài khoản thanh toán' })
+    return res.status(200).json({ message: 'Cannot find payment account' })
   }
   const destinationAccount = await accountModel.findOne({ uuid: data.destination_account_uuid }, accountViewModel)
   if (!destinationAccount) {
-    return res.status(200).json({ message: 'Không tìm thấy tài khoản người nhận' })
+    return res.status(200).json({ message: 'Cannot find receiver account' })
   }
   const transactionCode = 'SWEN' + otpGenerator.generate(15, { upperCaseAlphabets: false, specialChars: false, lowerCaseAlphabets: false })
   const transferData = {
@@ -171,7 +184,7 @@ router.post('/', async (req, res) => {
   } catch (err) {
     console.log('err', err)
   }
-  return res.status(200).json(transactionTransfer[0])
+  return res.status(201).json(transactionTransfer[0])
 })
 
 export default router
